separated sections into components for easier traversing; fiddled with styling for About section
