+++ src/components/About.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import '../index.css';
+
+const About = () => {
+    return (
+        <div id="about" className="About">
+            {/* Who am I? Example here */}
+            <h1 className="Head-Line">Hi, I'm Jack Mitchell!</h1>
+            <p>I'm an aspiring Full Stack Software Developer located in Austin, Texas.</p>
+            <p>Learning is my passion!</p>
+        </div>
+    );
+}
+
+export default About;
